requirements.txt
Flask
flask-cors
PyMuPDF
google-generativeai
dotenv